tests/wpt/web-platform-tests/fetch/sec-metadata/portal.tentative.https.sub.html

<!DOCTYPE html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=/fetch/sec-metadata/resources/helper.js></script>
<script src=/common/utils.js></script>
<body>
<script>
  const USER = true;
  const FORCED = false;

  function create_test(host, expectations) {
    async_test(t => {
      let p = document.createElement('portal');
      p.addEventListener('message', t.step_func(e => {
        assert_header_equals(e.data, expectations);
        t.done();
      }));

      let url = `https://${host}/fetch/sec-metadata/resources/post-to-owner.py`;
      p.src = url;
      document.body.appendChild(p);
    }, `{{host}} -> ${host} portal`);
  }

  create_test("{{host}}:{{ports[https][0]}}", {
    // TODO(mkwst): 'document' seems right, I guess? Perhaps a portal-specific destination would be better?
    "dest": "document",
    "site": "same-origin",
    "user": "",
    "mode": "nested-navigate"
  });

  create_test("{{hosts[][www]}}:{{ports[https][0]}}", {
    "dest": "document",
    "site": "same-site",
    "user": "",
    "mode": "nested-navigate"
  });

  create_test("{{hosts[alt][www]}}:{{ports[https][0]}}", {
    "dest": "document",
    "site": "cross-site",
    "user": "",
    "mode": "nested-navigate"
  });
</script>
